requirements.txt
Django>=5.2
djangorestframework>=3.20   
djangorestframework-simplejwt
django-cors-headers
pandas>=2.1
numpy>=1.28
scikit-learn>=1.3
joblib>=1.3
gunicorn>=21.2
            